refactor: extract bass pitch diagnostics
